+++ src/algebra/sub.rs
@@ -16,12 +16,12 @@ impl<N> Sub<Signal<N>> for Signal<N>
     }
 }
 
-impl<N> Sub<f64> for Signal<N>
-    where N: Sub<f64, Output = N> 
+impl<N> Sub<f32> for Signal<N>
+    where N: Sub<f32, Output = N> 
 {
     type Output = Signal<N>;
 
-    fn sub(self, rhs: f64) -> Self::Output {
+    fn sub(self, rhs: f32) -> Self::Output {
         Signal {
             value: self.value - rhs,
             domain: (self.domain.start()-rhs)..=(self.domain.end()-rhs)
@@ -29,9 +29,9 @@ impl<N> Sub<f64> for Signal<N>
     }
 }
 
-impl<N, __> Sub<Signal<N>> for f64
-    where f64: Sub<N, Output = N>,
-    f64: Sub<__>,
+impl<N, __> Sub<Signal<N>> for f32
+    where f32: Sub<N, Output = N>,
+    f32: Sub<__>,
     N: IsSame<This = __>,
 {
     type Output = Signal<N>;
